requirements.txt
streamlit>=1.52.0
pandas>=2.0.0
requests>=2.32.0
python-dotenv>=1.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
google-play-scraper>=1.2.0
PyJWT>=2.8.0

